LayoutTests/imported/w3c/web-platform-tests/css/css-anchor-position/anchor-getComputedStyle-003-expected.txt

anchor1anchor2target1target2

FAIL getComputedStyle() should return and absolutize the first @try rule style for target1 assert_equals: expected "100px" but got "999999px"
FAIL getComputedStyle() should return and absolutize the second @try rule style for target2 assert_equals: expected "200px" but got "0px"
